requirements.txt
Django>=5,<6
djangorestframework>=3.15,<4
